curasense-ml/requirements.txt
transformers
torch
crewai
crewai-tools
langchain
langchain-google-genai
langchain_community
tavily-python
python-dotenv
groq
fastapi
uvicorn
python-multipart
PyPDF2
pdfplumber
accelerate
pyyaml
huggingface_hub
